BACKPORT: fuse: enable caching of symlinks
FUSE file reads are cached in the page cache, but symlink reads are
not. This patch enables FUSE READLINK operations to be cached which
can improve performance of some FUSE workloads.

In particular, I'm working on a FUSE filesystem for access to source
code and discovered that about a 10% improvement to build times is
achieved with this patch (there are a lot of symlinks in the source
tree).

Change-Id: Ib62a80d4156ca40d0c4ddb8c34bfff1e1c920659
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Santhosh <santhosh.user.why.red@gmail.com>