UPSTREAM: afs: Add file locking tracepoints
Add two tracepoints for monitoring AFS file locking.  Firstly, add one that
follows the operational part:

    echo 1 >/sys/kernel/debug/tracing/events/afs/afs_flock_op/enable

And add a second that more follows the event-driven part:

    echo 1 >/sys/kernel/debug/tracing/events/afs/afs_flock_ev/enable

Individual file_lock structs seen by afs are tagged with debugging IDs that
are displayed in the trace log to make it easier to see what's going on,
especially as setting the first lock always seems to involve copying the
file_lock twice.

Change-Id: Ib581ecfd7d34508c9ddcf1b2161b5250a13b207a
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Santhosh <santhosh.user.why.red@gmail.com>