UPSTREAM: ANDROID: dma-buf: Fix error path on system heaps use of the page pool
Daniel Mentz relayed to me that Janghyuck Kim noted that
dma_buf_page_pool_create() returns NULL on error, so the
IS_ERR() check on the return value will never be true.

This patch fixes up the checking so we look for NULL instead and
properly return ENOMEM in that case.

Cc: Daniel Mentz <danielmentz@google.com>
Reported-by: Janghyuck Kim <janghyuck.kim@samsung.com>
Signed-off-by: John Stultz <john.stultz@linaro.org>
Fixes: 23762f02e172 ("ANDROID: dma-buf: system_heap: Add pagepool support to system heap")
Change-Id: Ia428e262cda5cc6bd39a3fbcddade7877473a563
Signed-off-by: Santhosh <santhosh.user.why.red@gmail.com>