regulator: cpr3-regulator: Add support for overvolting SoC
Since we already added support for undervolting, we need to adjust the logic for applying the final voltage.

To ensure the final voltage is accurately evaluated from both the reduction and addition,
we need to consider both adjustments together. The final voltage for each corner should be calculated as follows:

1. Subtract the custom voltage reduction from the reference voltage.
2. Add the custom voltage increase to the resulting voltage.

Signed-off-by: Santhosh <santhosh.user.why.red@gmail.com>